Partial revert to relative file names to allow exporting to html and tex directories separately.
The export-to-backend function now takes a relative path and exports to the export-html-dir or export-tex-dir depending on the backend.

I realised there's not point in having a build directory and an export directory inside the build directory. All I need is an export directory. Thus the build_dir option has been removed.

There's also no need to create a json build spec since I build the YAML front matter on export and that contains everything necessary for post-processing. So the build_json config option has also been removed.