Create export now copies site files rather than symlinking.
It's more flexible to copy files instead of symlinking. For example, now I can copy other files into the export site folder for the export process. Previously since that was symlinked to the site folder, any files copied there would be copied to the site which is not desirable - the files in question were not in the site folder for a reason!

Also created a key binding for ease of rebuilding the export environment whenever the site files (or any other relevant files) are changed.